requirements.txt
# Screen Capture Tool Dependencies
# Python 3.8+ required

PyQt5>=5.15.0
Pillow>=9.0.0
